Move summarizeSampleSetDist to simpleValue
